requirements.txt
aiogram==3.3.0
aiosqlite==0.19.0
httpx
openai
python-dotenv
